requirements.txt
opencv-python==4.8.0.74
Pillow==9.5.0
streamlit==1.25.0
ultralytics==8.0.149
